Moved fixtures to conftest, added session scope tmp_dir
For the autoML models to be useable for the entire session without
training, they require a session scoped tmp_dir. I tried to figure out
how to make the tmp_dir more dynamic but documentation seems to imply
that the scope is set at *function definition*, not on function call.
This means either call the _tmp_dir and manually clean up or just
duplicate the tmp_dir function but aptly named for session scope. It's a
bit ugly but couldn't find an alternative.
